# StreamSyncBackend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: StreamSync API
  version: 1.0.0
  description: Real-time collaborative workspace backend. Workspaces, projects, documents,
    tasks, activity and AI assistance.
paths:
...
  /api/workspaces/:
    get:
      operationId: workspaces_list
      description: Only workspaces where the caller holds an active membership. Pending
        invitations are not included — see `/api/workspaces/invitations/`.
      summary: List my workspaces
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      tags:
      - workspaces
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedWorkspaceList'
          description: ''
    post:
      operationId: workspaces_create
      description: The creator becomes its owner in the same transaction.
      summary: Create a workspace
      tags:
      - workspaces
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceCreateRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
          description: ''
  /api/workspaces/invitations/:
    get:
      operationId: workspaces_my_invitations
      description: 'Names the workspace and the inviter: this is what the recipient
        decides on, and it is not reachable from anywhere else — a workspace you have
        not joined is absent from GET /api/workspaces/.'
      summary: My pending invitations
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      tags:
      - workspaces
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPendingInvitationList'
          description: ''
  /api/workspaces/{workspace_id}/:
    get:
      operationId: workspaces_retrieve
      description: 404 for a workspace the caller is not a member of.
      summary: Workspace detail
      parameters:
      - in: path
        name: workspace_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - workspaces
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
          description: ''
        '404':
          description: Not found.
    patch:
      operationId: workspaces_update
      description: Owner only. The slug is fixed at creation and never changes.
      summary: Rename a workspace
      parameters:
      - in: path
        name: workspace_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - workspaces
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedWorkspaceUpdateRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
          description: ''
        '403':
          description: Not the owner.
    delete:
      operationId: workspaces_destroy
      description: Owner only. Removes every membership with it.
      summary: Delete a workspace
      parameters:
      - in: path
        name: workspace_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - workspaces
      security:
      - jwtAuth: []
      responses:
        '204':
          description: Deleted.
  /api/workspaces/{workspace_id}/members/:
    get:
      operationId: workspaces_members_list
      description: Active members and pending invitations together, distinguished
        by `status`. Any member of the workspace may read this.
      summary: List members
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - in: path
        name: workspace_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - workspaces
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedWorkspaceMemberList'
          description: ''
  /api/workspaces/{workspace_id}/members/{membership_id}/:
    patch:
      operationId: workspaces_member_update
      description: Owner only. The owner's own role cannot be changed here.
      summary: Change a member's role
      parameters:
      - in: path
        name: membership_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: workspace_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - workspaces
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedUpdateMemberRoleRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceMember'
          description: ''
    delete:
      operationId: workspaces_member_remove
      description: The owner may remove anyone; any member may remove themselves.
        The owner cannot be removed — transfer ownership first.
      summary: Remove a member or leave
      parameters:
      - in: path
        name: membership_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: workspace_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - workspaces
      security:
      - jwtAuth: []
      responses:
        '204':
          description: Removed.
  /api/workspaces/{workspace_id}/invitations/:
    post:
      operationId: workspaces_invite
      description: Owner only. Creates a membership in the `invited` state. The invitee
        must already have a StreamSync account — emailing a signup link needs the
        mail pipeline from Milestone 8.
      summary: Invite a member
      parameters:
      - in: path
        name: workspace_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - workspaces
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteMemberRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceMember'
          description: ''
        '400':
          description: Unknown email, or owner role requested.
        '403':
          description: Not the workspace owner.
        '409':
          description: Already a member or already invited.
  /api/workspaces/{workspace_id}/invitations/accept/:
    post:
      operationId: workspaces_invitation_accept
      description: |-
        Accept an invitation.

        Cannot use `visible_workspaces`, which by definition excludes workspaces
        the caller has not joined. The invitation itself is the authorization: the
        lookup is scoped to an INVITED membership belonging to the caller, so
        nobody can accept an invitation addressed to someone else.
      summary: Accept an invitation
      parameters:
      - in: path
        name: workspace_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - workspaces
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceMember'
          description: ''
        '400':
          description: No pending invitation.
        '404':
          description: No such workspace.
...
components:
  schemas:
...
    AssignableWorkspaceRole:
      enum:
      - editor
      - viewer
      type: string
      description: |-
        * `editor` - Editor
        * `viewer` - Viewer
...
    InviteMemberRequest:
      type: object
      description: Invitation input. The inviter comes from the session, never the
        body.
      properties:
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 254
        role:
          allOf:
          - $ref: '#/components/schemas/AssignableWorkspaceRole'
          default: viewer
      required:
      - email
...
    PaginatedPendingInvitationList:
      type: object
      required:
      - count
      - page
      - page_size
      - total_pages
      - results
      properties:
        count:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 25
        total_pages:
          type: integer
          example: 2
        next:
          type: string
          nullable: true
          format: uri
        previous:
          type: string
          nullable: true
          format: uri
        results:
          type: array
          items:
            $ref: '#/components/schemas/PendingInvitation'
...
    PaginatedWorkspaceList:
      type: object
      required:
      - count
      - page
      - page_size
      - total_pages
      - results
      properties:
        count:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 25
        total_pages:
          type: integer
          example: 2
        next:
          type: string
          nullable: true
          format: uri
        previous:
          type: string
          nullable: true
          format: uri
        results:
          type: array
          items:
            $ref: '#/components/schemas/Workspace'
    PaginatedWorkspaceMemberList:
      type: object
      required:
      - count
      - page
      - page_size
      - total_pages
      - results
      properties:
        count:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 25
        total_pages:
          type: integer
          example: 2
        next:
          type: string
          nullable: true
          format: uri
        previous:
          type: string
          nullable: true
          format: uri
        results:
          type: array
          items:
            $ref: '#/components/schemas/WorkspaceMember'
...
    PatchedUpdateMemberRoleRequest:
      type: object
      description: Role change input. Owner is excluded — ownership is transferred.
      properties:
        role:
          $ref: '#/components/schemas/AssignableWorkspaceRole'
    PatchedWorkspaceUpdateRequest:
      type: object
      description: |-
        Partial update input.

        `slug` is absent by design: it is assigned once and never changes, so that
        existing links keep working after a rename.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          nullable: true
          maxLength: 500
    PendingInvitation:
      type: object
      description: |-
        An invitation from the invitee's side.

        Distinct from `WorkspaceMemberSerializer`, which describes somebody *within*
        a workspace you are already looking at and therefore has no reason to name
        it. Here the workspace is the whole point: the recipient is deciding whether
        to join, and "you have been invited as an editor" without saying by whom, or
        to what, is not a decision anyone can make.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        workspace_id:
          type: string
          format: uuid
          readOnly: true
        workspace_name:
          type: string
          readOnly: true
        workspace_slug:
          type: string
          readOnly: true
        role:
          allOf:
          - $ref: '#/components/schemas/WorkspaceRole'
          readOnly: true
        invited_by:
          allOf:
          - $ref: '#/components/schemas/Collaborator'
          readOnly: true
        invited_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - id
      - invited_at
      - invited_by
      - role
      - workspace_id
      - workspace_name
      - workspace_slug
...
    User:
      type: object
      description: |-
        The public shape of a user.

        Explicit `fields` rather than `exclude`: a field added to the model must
        never appear in an API response merely because someone forgot to hide it.
        That is how password hashes and internal flags leak.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          readOnly: true
          title: Full name
          description: Display name shown to collaborators.
        email:
          type: string
          format: email
          readOnly: true
          title: Email address
          description: Used to sign in. Case-insensitively unique.
        avatar_url:
          type: string
          readOnly: true
        title:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - avatar_url
      - created_at
      - email
      - id
      - name
      - title
    Workspace:
      type: object
      description: |-
        A workspace as seen *by the requesting user*.

        `role` is not a property of the workspace — it is a property of the
        relationship between this workspace and whoever is asking. It is included
        here because every screen that renders a workspace also needs to know which
        actions to offer, and a second request per workspace to find that out would
        be wasteful.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          readOnly: true
        slug:
          type: string
          readOnly: true
          description: URL-safe identifier. Assigned at creation and never changes.
          pattern: ^[-a-zA-Z0-9_]+$
        description:
          type: string
          readOnly: true
        member_count:
          type: integer
          readOnly: true
        role:
          type: string
          nullable: true
          description: |-
            Read the annotation the view's queryset attached.

            Falls back to a lookup only when the serializer is used outside that
            queryset, so the list endpoint stays at one query rather than one per
            workspace.
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - description
      - id
      - member_count
      - name
      - role
      - slug
    WorkspaceCreateRequest:
      type: object
      description: |-
        Creation input.

        A plain Serializer rather than a ModelSerializer so that `owner` and `slug`
        are simply not accepted from the client. Both are decided by the service —
        a caller who could set `owner` could create a workspace belonging to
        somebody else.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          nullable: true
          default: ''
          maxLength: 500
      required:
      - name
    WorkspaceMember:
      type: object
      description: One row in the members list — active members and pending invitations.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        user:
          allOf:
          - $ref: '#/components/schemas/User'
          readOnly: true
        role:
          allOf:
          - $ref: '#/components/schemas/WorkspaceRole'
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/MembershipStatus'
          readOnly: true
        joined_at:
          description: |-
            Never null, so the client can always render a date.

            An invited person has not joined yet, so the date shown is when they
            were invited. `status` is what distinguishes the two, and the UI reads
            that rather than inferring from the timestamp.
          readOnly: true
      required:
      - id
      - joined_at
      - role
      - status
      - user
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT